+++ plugin/src/main/java/io/brachu/docker/compose/plugin/Config.java
@@ -3,12 +3,13 @@ package io.brachu.docker.compose.plugin;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Map;
+import java.util.Optional;
 
 import javax.annotation.Nullable;
 
 import org.apache.commons.lang3.StringUtils;
 
-final class CommonConfig {
+final class Config {
 
     private final String executablePath;
 
@@ -18,15 +19,18 @@ final class CommonConfig {
 
     private final Map<String, String> env;
 
+    private final WaitConfig wait;
+
     private final boolean skip;
 
-    CommonConfig(@Nullable String executablePath, String projectBasedir, String file, @Nullable String projectName, @Nullable Map<String, String> env,
-                 boolean skip) {
+    Config(@Nullable String executablePath, String projectBasedir, String file, @Nullable String projectName, @Nullable Map<String, String> env,
+           @Nullable WaitConfig wait, boolean skip) {
 
         this.executablePath = StringUtils.trimToNull(executablePath);
         this.file = prepareFilePath(projectBasedir, file);
         this.projectName = StringUtils.trimToNull(projectName);
         this.env = env;
+        this.wait = Optional.ofNullable(wait).orElseGet(WaitConfig::new);
         this.skip = skip;
     }
 
@@ -44,6 +48,10 @@ final class CommonConfig {
 
     Map<String, String> getEnv() {
         return env;
+    }
+
+    WaitConfig getWait() {
+        return wait;
     }
 
     boolean shouldExecute() {