+++ plugin/src/main/java/io/brachu/docker/compose/plugin/ConfigValidator.java
@@ -4,13 +4,20 @@ import java.util.concurrent.TimeUnit;
 
 import org.apache.maven.plugin.MojoFailureException;
 
-final class UpConfigValidator {
+final class ConfigValidator {
 
-    static void validate(UpConfig config) throws MojoFailureException {
+    static void validate(Config config) throws MojoFailureException {
+        validateFile(config);
         validateWait(config);
     }
 
-    private static void validateWait(UpConfig config) throws MojoFailureException {
+    private static void validateFile(Config config) throws MojoFailureException {
+        if (config.getFile() == null) {
+            throw new MojoFailureException("Path to docker-compose.yml file must be set.");
+        }
+    }
+
+    private static void validateWait(Config config) throws MojoFailureException {
         WaitConfig wait = config.getWait();
         if (wait != null) {
             if (wait.getValue() < 1) {