+++ plugin/src/main/java/io/brachu/docker/compose/plugin/DownMojo.java
@@ -3,7 +3,6 @@ package io.brachu.docker.compose.plugin;
 import io.brachu.johann.DockerCompose;
 import io.brachu.johann.DownConfig;
 import io.brachu.johann.exception.JohannException;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
@@ -39,7 +38,7 @@ public final class DownMojo extends AbstractDockerComposeMojo {
 
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
-        CommonConfig config = getCommonConfig();
+        Config config = getConfig();
         if (config.shouldExecute()) {
             DockerCompose compose = dockerCompose(config);
             clearProperties(compose);
@@ -65,32 +64,6 @@ public final class DownMojo extends AbstractDockerComposeMojo {
                 .withRemoveVolumes(removeVolumes)
                 .withRemoveOrphans(removeOrphans)
                 .withTimeoutSeconds(downTimeoutSeconds);
-    }
-
-    private void clearProperties(DockerCompose compose) {
-        clearProjectProperties();
-        clearFailsafeArgLine(compose);
-        clearSystemProperties();
-    }
-
-    private void clearProjectProperties() {
-        project.getProperties().remove(PROJECT_NAME_PROPERTY);
-    }
-
-    private void clearFailsafeArgLine(DockerCompose compose) {
-        String failsafeArgLine = project.getProperties().getProperty(FAILSAFE_ARGLINE_PROPERTY);
-        if (failsafeArgLine != null) {
-            failsafeArgLine = StringUtils.trimToNull(failsafeArgLine.replace(constructFailsafeArgLine(compose), ""));
-            if (failsafeArgLine != null) {
-                project.getProperties().setProperty(FAILSAFE_ARGLINE_PROPERTY, failsafeArgLine);
-            } else {
-                project.getProperties().remove(FAILSAFE_ARGLINE_PROPERTY);
-            }
-        }
-    }
-
-    private void clearSystemProperties() {
-        System.clearProperty(PROJECT_NAME_PROPERTY);
     }
 
 }